Organize server logic
